restructure the directory structures of the handlers and utils. Moved constants to its own file
