+++ praetorian_cli/handlers/actions/delete.py
@@ -1,7 +1,7 @@
 import click
 
-from praetorian_cli.handlers.chariot import chariot
-from praetorian_cli.handlers.cli_decorators import cli_handler
+from praetorian_cli.handlers.products import chariot
+from praetorian_cli.handlers.utils.cli_decorators import cli_handler
 
 
 @chariot.group()
