+++ praetorian_cli/handlers/actions/get.py
@@ -3,8 +3,8 @@ import json
 
 import click
 
-from praetorian_cli.handlers.chariot import chariot
-from praetorian_cli.handlers.cli_decorators import cli_handler, scripts
+from praetorian_cli.handlers.products import chariot
+from praetorian_cli.handlers.utils.cli_decorators import cli_handler, scripts
 
 
 @chariot.group()
