+++ praetorian_cli/handlers/actions/update.py
@@ -1,8 +1,8 @@
 import click
 
-from praetorian_cli.handlers.chariot import chariot
-from praetorian_cli.handlers.utils import Status
-from praetorian_cli.handlers.cli_decorators import cli_handler, status_options
+from praetorian_cli.handlers.products import chariot
+from praetorian_cli.sdk.constants import Status
+from praetorian_cli.handlers.utils.cli_decorators import cli_handler, status_options
 
 
 @chariot.group()
